Add silo data, annotation, and link
